requirements.txt
mysql-connector
FastAPI
uvicorn
python-dotenv
pydantic
requests
beautifulsoup4
sqlalchemy
pymysql
